Ref: helpers -> common helpers. Extracting OpenAPI helpers.
