+++ src/common-helpers.ts
