+++ src/endpoints-factory.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 import { ApiResponse } from "./api-response";
 import { Endpoint, Handler } from "./endpoint";
-import { FlatObject, IOSchema, Merge } from "./helpers";
+import { FlatObject, IOSchema, Merge } from "./common-helpers";
 import { Method, MethodsDefinition } from "./method";
 import { MiddlewareDefinition } from "./middleware";
 import { mimeJson, mimeMultipart } from "./mime";
