+++ src/result-handler.ts
@@ -8,7 +8,7 @@ import {
   getStatusCodeFromError,
   IOSchema,
   markOutput,
-} from "./helpers";
+} from "./common-helpers";
 import { getMeta, withMeta } from "./metadata";
 
 interface LastResortHandlerParams {
